+++ include/swift/SILOptimizer/Utils/InstOptUtils.h
@@ -423,14 +423,6 @@ bool canReplaceLoadSequence(SILInstruction *inst);
 /// starting with the innermost struct_element_addr
 void replaceLoadSequence(SILInstruction *inst, SILValue value);
 
-/// Do we have enough information to determine all callees that could
-/// be reached by calling the function represented by Decl?
-bool calleesAreStaticallyKnowable(SILModule &module, SILDeclRef decl);
-
-/// Do we have enough information to determine all callees that could
-/// be reached by calling the function represented by Decl?
-bool calleesAreStaticallyKnowable(SILModule &module, ValueDecl *vd);
-
 // Attempt to get the instance for , whose static type is the same as
 // its exact dynamic type, returning a null SILValue() if we cannot find it.
 // The information that a static type is the same as the exact dynamic,
